backend/requirements.txt
# 基础异步/网络依赖
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
multidict==6.0.5
requests==2.32.3
sniffio==1.3.1
urllib3==2.2.2
yarl==1.9.4

# 数据处理/解析依赖
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
chardet==5.2.0
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fonttools==4.49.0
humanfriendly==10.0
importlib-resources==6.1.1
iopath==0.1.10
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
json-repair==0.25.2
kiwisolver==1.4.5
langdetect==1.0.9
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.7.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
orjson==3.9.15
packaging==23.2
pandas==2.2.0
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pikepdf==8.11.0
pillow==9.5.0
pillow_heif==0.15.0
portalocker==2.8.2
pycparser==2.21
pyparsing==3.0.9
pypdf==4.0.1
pypdfium2==4.27.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
scipy==1.10.1
shapely==2.0.3
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.28
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
wikipedia==1.4.0
wrapt==1.16.0
youtube-transcript-api==0.6.2
zipp==3.17.0

# 安全/加密依赖
cryptography==42.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
rsa==4.9

# 云服务依赖
boto3==1.34.140
minio==7.2.7
botocore==1.34.140
s3transfer==0.10.1
google-api-core==2.18.0
google-auth==2.29.0
google_auth_oauthlib==1.2.0
google-cloud-aiplatform==1.58.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-logging==3.10.0
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-ai-generativelanguage==0.6.6
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
proto-plus==1.23.0
protobuf==3.20.3
psutil==6.0.0

# LLM/向量/知识库核心依赖
cachetools==5.3.3
fastapi==0.111.0
fastapi-health==0.4.0
greenlet==3.0.3
huggingface-hub==0.22.2
langchain==0.2.6
langchain-aws==0.1.9
langchain-anthropic==0.1.19
langchain-community==0.2.6
langchain-core==0.2.10
langchain-experimental==0.0.62
langchain-fireworks==0.1.4
langchain-google-genai==1.0.7
langchain-google-vertexai==1.0.6
langchain-groq==0.1.6
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langsmith==0.1.83
langserve==0.2.2
layoutparser==0.3.4
openai==1.35.10
opencv-python==4.8.0.76
pydantic==2.8.2
pydantic_core==2.20.1
pycocotools==2.0.7
sentence-transformers==2.7.0
safetensors==0.4.1
starlette==0.37.2
sse-starlette==2.1.2
starlette-session==0.4.3
tiktoken==0.7.0
timm==0.9.12
tokenizers==0.19.1
transformers==4.42.3
unstructured==0.14.9
unstructured-client==0.23.8
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
unstructured[all-docs]==0.14.9

# 部署/服务依赖
coloredlogs==15.0.1
effdet==0.4.1
emoji==2.10.1
gunicorn==22.0.0
uvicorn==0.30.1

# 第三方模型/工具依赖
PyMuPDF==1.24.5
modelscope==1.15.0
volcengine-python-sdk[ark]
types-protobuf
types-requests
install==1.3.5
zhipuai>=2.1.0